+++ Database/Models/Map.cs
@@ -2,9 +2,9 @@
 
 namespace Database.Models
 {
-    public class Map : IHaveId<long>
+    public class Map : IHaveId<int>
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public string Name { get; set; }
