+++ Database/Models/Migration.cs
@@ -2,9 +2,9 @@
 
 namespace Database.Models
 {
-    public class Migration : IHaveId<long>
+    public class Migration : IHaveId<int>
     {
-        public long Id { get; private set; }
+        public int Id { get; private set; }
 
         public string Name { get; set; }
     }
