+++ LiveView.WebApi/Dto/ServerDto.cs
@@ -3,9 +3,9 @@ using System.Text.Json.Serialization;
 
 namespace LiveView.WebApi.Dto
 {
-    public record ServerDto : IHaveIdWithSetter<long>
+    public record ServerDto : IHaveIdWithSetter<int>
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public string? IpAddress { get; set; }
