+++ LiveView.WebApi/Dto/UserDto.cs
@@ -3,9 +3,9 @@ using System.Text.Json.Serialization;
 
 namespace LiveView.WebApi.Dto
 {
-    public record UserDto : IHaveIdWithSetter<long>
+    public record UserDto : IHaveIdWithSetter<int>
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public string Username { get; set; }
