+++ LiveView.WebApi/Dto/UserEventDto.cs
@@ -3,9 +3,9 @@ using System.Text.Json.Serialization;
 
 namespace LiveView.WebApi.Dto
 {
-    public record UserEventDto : IHaveIdWithSetter<long>
+    public record UserEventDto : IHaveIdWithSetter<int>
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public string Note { get; set; }
