+++ LiveView/Services/UniqueIdGenerator.cs
@@ -6,15 +6,15 @@ namespace LiveView.Services
 {
     public static class UniqueIdGenerator
     {
-        public static long Get(Enum @enum)
+        public static int Get(Enum @enum)
         {
             return Get(String.Concat(@enum.GetType().Name, @enum));
         }
 
-        public static long Get(string input)
+        public static int Get(string input)
         {
             var hashBytes = SHA256.Create().ComputeHash(Encoding.UTF8.GetBytes(input));
-            return BitConverter.ToInt64(hashBytes, 0);
+            return BitConverter.ToInt32(hashBytes, 0);
         }
     }
 }
